requirements.txt
filelock==3.13.1
fsspec==2024.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sympy==1.13.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
typing_extensions==4.9.0
tzdata==2024.2